+++ test/test_control/test_main.cpp
@@ -400,6 +400,7 @@ void test_actual_loop_bounds_and_stop_with_rx_backlog() {
 }
 
 void runFuzzTests();
+void runConsoleTests();
 
 int main() {
   UNITY_BEGIN();
@@ -422,5 +423,6 @@ int main() {
   RUN_TEST(test_logging_budget_backpressure_and_whole_line_drops);
   RUN_TEST(test_actual_loop_bounds_and_stop_with_rx_backlog);
   runFuzzTests();
+  runConsoleTests();
   return UNITY_END();
 }